restructure logs
