+++ platform/src/bridge.ts
@@ -2,11 +2,8 @@
 import { $ } from "bun";
 import { interpreter as py } from "node-calls-python";
 import path from "node:path";
-import readline from "node:readline";
-import fs from "node:fs";
-import { rm } from "node:fs/promises";
-import crypto from "node:crypto";
-import { spawn } from "node:child_process";
+
+import setupLogger from "./util/streaming-logs";
 
 // Use the major.minor python version to find the local poetry venv
 // see https://github.com/OpenFn/gen/issues/45
@@ -20,53 +17,6 @@ if (allowReimport) {
   );
 }
 
-/**
- * This nasty little function sets up logfile and a watch
- * to capture the srdout from the python process
- * Python's logger is configured to write to this file,
- * which will be watched and then re-directed
- *
- * TODO: I'd like this to just be a promise, it'll be a bit slicker,
- * But I'm still sketching it out really
- *
- */
-const setupLogger = (onLog = (_l: string) => {}, onComplete = () => {}) => {
-  const id = crypto.randomUUID();
-
-  const logfile = path.resolve(`tmp/logs/${id}.log`);
-  Bun.write(logfile, "");
-  console.log("node reading logs from", logfile);
-
-  // attempt #1: use bun shell
-  // sadly bun shell doesn't support streams so I cant do this:
-  // $`tail -f ${logfile}`;
-
-  // attempt #2: use child process
-  // This is me trying not to spawn a process to run python. hmm.
-  const child = spawn("tail", ["-f", logfile]);
-  const rl = readline.createInterface({
-    input: child.stdout,
-    crlfDelay: Infinity,
-  });
-  rl.on("line", (line) => {
-    if (line === "***END***") {
-      destroy();
-      onComplete();
-    } else {
-      onLog(line);
-    }
-  });
-
-  // kill the child process and remove the log file
-  const destroy = async () => {
-    child.kill();
-
-    await rm(logfile);
-  };
-
-  return { logfile, destroy };
-};
-
 // TODO I need to make this blocking so that only one thing runs at once
 // OR I drop workerpool onto it
 export const run = async (
@@ -75,7 +25,7 @@ export const run = async (
   onLog?: (str: string) => void
 ) => {
   try {
-    // poetry should be configured to use a vnv in the local filesystem
+    // poetry should be configured to use a venv in the local filesystem
     // This makes it really easy to tell node-calls-python about the right env!
     // IMPORTANT: if the python version changes, this path needs to be manually updated!
     py.addImportPath(
@@ -96,7 +46,7 @@ export const run = async (
         resolve(result);
       };
 
-      const { logfile, destroy } = setupLogger(onLog, onComplete);
+      const { logfile, delimiter, destroy } = setupLogger(onLog, onComplete);
 
       // import from a top level entry point
       const pymodule = await py.import(
@@ -105,7 +55,12 @@ export const run = async (
       );
 
       try {
-        result = await py.call(pymodule, "main", [scriptName, args, logfile]);
+        result = await py.call(pymodule, "main", [
+          scriptName,
+          args,
+          logfile,
+          delimiter,
+        ]);
       } catch (e) {
         // Note that the error coming out will be a string with no stack trace :(
         console.log(e);
